+++ kazoo/run.sh
@@ -3,8 +3,16 @@ FLAGS=${1:-"-td"}
 COMMIT=${2:-""}
 NETWORK=${NETWORK:-"kazoo"}
 NAME=kazoo.$NETWORK
-docker stop -t 1 $NAME
-docker rm -f $NAME
+
+if [ -n "$(docker ps -aq -f name=$NAME)" ]
+then
+   echo -n "stopping: "
+   docker stop -t 1 $NAME
+   echo -n "removing: "
+   docker rm -f $NAME
+fi
+echo -n "starting: $NAME "
+
 docker run $FLAGS \
 	--net $NETWORK \
 	-h $NAME \
